test(parser): unit tests of parser api `setFilterEleTypes` for both eam/fs and eam/alloy potential
In this commit, we add unit tests for parser api `setFilterEleTypes` by calling api
setFilterEleTypes and comparing the parsing results (potential tables) with the randomly generated
ones in fixture `EamPotAlloyTestFixture`. To compare the potential tables, we adjust the original
table generation method in `EamPotAlloyTestFixture` to make sure the data in each table are unique.